requirements.txt
alembic==1.8.1
appdirs==1.4.4
APScheduler==3.9.1.post1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
bcrypt==4.0.1
click==8.1.3
decorator==5.1.1
dnspython==2.3.0
email-validator==1.3.0
exceptiongroup==1.1.0
executing==1.2.0
Faker==15.3.4
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
gnureadline==8.1.2
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
ipdb==0.13.11
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.36
protobuf==3.20.3
ptpython==3.0.22
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pydantic==1.10.4
Pygments==2.13.0
PyJWT==2.6.0
PyMySQL==1.0.2
pytest==7.2.0
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.4.2
six==1.16.0
SQLAlchemy==1.4.45
SQLAlchemy-Utils==0.39.0
stack-data==0.6.2
tomli==2.0.1
traitlets==5.7.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
wcwidth==0.2.5
Werkzeug==2.2.2
